backend/services/resume-tailor/requirements.txt
# Core dependencies for Resume Tailor CLI
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Optional: Keep these for future web service integration
# openai
# boto3
# fastapi
# uvicorn

